[AT] update: 파일명 변경(css-> scss , AlcoholType->AT)
