+++ components/AlocholType/ATCard/ATCard.tsx
@@ -1,4 +1,4 @@
-import styles from './AlcoholTypeInfo.module.css';
+import styles from './ATCard.module.scss';
 
 interface AlcoholCardTypeProps {
     name: string;
@@ -11,13 +11,19 @@ interface AlcoholCardTypeProps {
     hashtags?: string[];
 }
 
-const AlcoholTypeInfo: React.FC<AlcoholCardTypeProps> = ({ ...info }: AlcoholCardTypeProps) => {
+interface AlcoholCardProps extends AlcoholCardTypeProps{
+    onClick: ()=>void;
+    type: 'sm' | 'md' | 'lg';
+}
+
+
+const ATCard: React.FC<AlcoholCardProps> = ({ type, onClick, ...info }: AlcoholCardProps) => {
     function changeNumberToMoney (num: number) {
         return num.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
     }
 
     return (
-        <section className={styles.container}>
+        <div className={styles.card} onClick={onClick}>
             <div className={styles.alcoholImageBox} >
                 <img className={styles.alcoholImage} src={info.image}></img>
             </div>
@@ -25,18 +31,14 @@ const AlcoholTypeInfo: React.FC<AlcoholCardTypeProps> = ({ ...info }: AlcoholCar
                 <p className={styles.title}>{info.name}</p>
                 <p className={styles.basic}>{info.category}</p>
                 <p className={styles.basic}>{info.degree}%</p>
-                <p className={styles.basic}>{info.drinkSize}ml
-                    (<span>
-                        {info.minPrice && changeNumberToMoney(info.minPrice)}₩~ {info.maxPrice && changeNumberToMoney(info.maxPrice)}₩
-                    </span>)
-                </p>
+                <p className={styles.basic}>{info.minPrice && changeNumberToMoney(info.minPrice)} ~ {info.maxPrice && changeNumberToMoney(info.maxPrice)}</p>
                 <div className={styles.hashtagBox}>
-                    {info.hashtags && info.hashtags.map((v) => (<div className={styles.hashtag}># {v}</div>))}
+                    {info.hashtags && info.hashtags.map((v) => (<div key={v} className={styles.hashtag}># {v}</div>))}
                 </div>
             </div>
-        </section>
+        </div>
     )
 
 };
 
-export default AlcoholTypeInfo;
+export default ATCard;
