+++ components/AlocholType/ATInfo/ATInfo.tsx
@@ -1,4 +1,4 @@
-import styles from './AlcoholCard.module.scss';
+import styles from './ATInfo.module.scss';
 
 interface AlcoholCardTypeProps {
     name: string;
@@ -11,19 +11,13 @@ interface AlcoholCardTypeProps {
     hashtags?: string[];
 }
 
-interface AlcoholCardProps extends AlcoholCardTypeProps{
-    onClick: ()=>void;
-    type: 'sm' | 'md' | 'lg';
-}
-
-
-const AlcoholCard: React.FC<AlcoholCardProps> = ({ type, onClick, ...info }: AlcoholCardProps) => {
+const AlcoholTypeInfo: React.FC<AlcoholCardTypeProps> = ({ ...info }: AlcoholCardTypeProps) => {
     function changeNumberToMoney (num: number) {
         return num.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
     }
 
     return (
-        <div className={styles.card} onClick={onClick}>
+        <section className={styles.container}>
             <div className={styles.alcoholImageBox} >
                 <img className={styles.alcoholImage} src={info.image}></img>
             </div>
@@ -31,14 +25,18 @@ const AlcoholCard: React.FC<AlcoholCardProps> = ({ type, onClick, ...info }: Alc
                 <p className={styles.title}>{info.name}</p>
                 <p className={styles.basic}>{info.category}</p>
                 <p className={styles.basic}>{info.degree}%</p>
-                <p className={styles.basic}>{info.minPrice && changeNumberToMoney(info.minPrice)} ~ {info.maxPrice && changeNumberToMoney(info.maxPrice)}</p>
+                <p className={styles.basic}>{info.drinkSize}ml
+                    (<span>
+                        {info.minPrice && changeNumberToMoney(info.minPrice)}₩~ {info.maxPrice && changeNumberToMoney(info.maxPrice)}₩
+                    </span>)
+                </p>
                 <div className={styles.hashtagBox}>
-                    {info.hashtags && info.hashtags.map((v) => (<div key={v} className={styles.hashtag}># {v}</div>))}
+                    {info.hashtags && info.hashtags.map((v) => (<div className={styles.hashtag}># {v}</div>))}
                 </div>
             </div>
-        </div>
+        </section>
     )
 
 };
 
-export default AlcoholCard;
+export default AlcoholTypeInfo;
