+++ components/AlocholType/ATResultList/ATResultList.module.scss
