+++ components/AlocholType/ATSimilarList/ATSimilarList.module.scss
