+++ components/AlocholType/ATSimilarList/index.tsx
@@ -1,3 +1,2 @@
 // import할 때 파일 명까지 적는 불편 + 실수의 위험성을 해결하고자 index파일로 모듈화할 내용들은 따로 기입해둡니다.
-export {default} from './AlcoholCard';
-
+export {default} from './ATSimilarList';
