--- components/AlocholType/AlcoholTypeInfo/AlcoholTypeInfo.module.css
@@ -1,59 +1,0 @@
-.container{
-    display: flex;
-    flex-direction: row;
-    gap: 5%;
-    padding: 0.5rem;
-    width: auto;
-    height: 35vh;
-}
-.alcoholImageBox{
-    width: 25%;
-    aspect-ratio: 5/8;
-    border-radius: 10px;
-    box-shadow: 2px 2px 4px  rgba(0,0,0,0.1);
-}
-.alcoholImage{
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-.alcoholInfoBox{
-    width: auto;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: left;
-    gap: 0.3rem;
-    font-size: 0.9rem;
-    overflow-x: hidden;
-}
-
-.title{
-    margin: 0;
-    font-size: 2rem;
-    font-weight: bold;
-    color: #000;
-    margin-bottom: 2rem;
-}
-.basic{
-    margin: 0;
-    font-size: 1rem;
-}
-.hashtagBox{
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-    flex-wrap: wrap;
-
-    padding-top: 0.5rem;
-    
-    gap: 10px;
-    font-size: 0.75rem;
-}
-.hashtag{
-    color: white;
-    background-color: #B7B7B7;
-    border-radius: 22px;
-    padding: 0.2rem 0.5rem;
-}
