+++ styles/AlcoholType/AlcoholType.module.scss
